Refactor frontend folder structure
