+++ cms/static-data.json
@@ -1,7 +1,7 @@
 {
   "homepage": {
     "createdAt": "2025-03-02T08:55:46.145Z",
-    "updatedAt": "2025-03-06T07:12:33.627Z",
+    "updatedAt": "2025-03-06T08:19:50.707Z",
     "globalType": "homepage",
     "heroBanner": {
       "headline": "Hi, I'm Hayden",
@@ -10,6 +10,133 @@
     "homeAbout": {
       "title": "About Me",
       "paragraph": "I'm a software developer and educator with a passion for creating a positive impact through programming, education, and user-centric design. I have a Bachelor's in Computer Science from Sonoma State University, am a Certified Scrum Master, and have worked in a variety of different roles including management, design, education, and quality assurance. In my free time, I enjoy developing products, learning new skills, reading, playing video games, and spending time outdoors with family and friends."
+    },
+    "homeExperience": {
+      "title": "Experience",
+      "experienceItems": [
+        {
+          "company": "HL Tutoring",
+          "titles": [
+            {
+              "title": "Founder",
+              "id": "67c954bdf3c2fbef53f39d39"
+            }
+          ],
+          "dates": {
+            "startDate": "2023",
+            "endDate": "Present",
+            "shortLabel": "Present"
+          },
+          "description": "Tutoring of a wide range of students with a focus on helping students gain a thorough understanding of technical topics and achieve a variety of educational goals. Instruct students in college level data structures and algorithms, AP Computer Science, Full Stack Web Development, and learning computer science fundamentals from the ground up.",
+          "technologies": [
+            "Python",
+            "Java",
+            "Javascript",
+            "React",
+            "Go",
+            "Express",
+            "Node.js",
+            "Next.js",
+            "Flask",
+            "Ollama",
+            "MongoDB",
+            "SQL",
+            "CSS",
+            "HTML",
+            "C++",
+            "C",
+            "Squarespace"
+          ],
+          "createdAt": "2025-03-06T07:55:33.072Z",
+          "updatedAt": "2025-03-06T08:15:00.211Z",
+          "id": "67c954f50e528355496e4293"
+        },
+        {
+          "company": "Nav Creative",
+          "titles": [
+            {
+              "title": "Director",
+              "id": "67c95547f3c2fbef53f39d3f"
+            },
+            {
+              "title": "Project Manager",
+              "id": "67c95551f3c2fbef53f39d41"
+            },
+            {
+              "title": "Web Designer",
+              "id": "67c95556f3c2fbef53f39d43"
+            }
+          ],
+          "dates": {
+            "startDate": "2019",
+            "endDate": "2023",
+            "shortLabel": "2023"
+          },
+          "description": "Designed and managed over 200 full website builds for small business clients under heavy workloads. Co-created the studio's new core process for web clients, fixing a critical revenue issue. Oversaw all high complexity and technical client lead consultations across their entire lifecycle, streamlining the onboarding of new clients and providing support during the building and offboarding phases of the process.",
+          "technologies": [
+            "Javascript",
+            "Less",
+            "JQuery",
+            "HTML",
+            "Python",
+            "Squarespace",
+            "Wordpress",
+            "Shopify"
+          ],
+          "createdAt": "2025-03-06T07:58:22.515Z",
+          "updatedAt": "2025-03-06T08:15:59.496Z",
+          "id": "67c9559e0e528355496e42d5"
+        },
+        {
+          "company": "Silicon Valley 4u",
+          "titles": [
+            {
+              "title": "Coding Instructor",
+              "id": "67c959daf3c2fbef53f39d49"
+            }
+          ],
+          "dates": {
+            "startDate": "May 2019",
+            "endDate": "Nov 2019",
+            "shortLabel": "2019"
+          },
+          "description": "Taught a variety of courses in Python, Scratch, and Blender 3D, adapting curriculum to different age groups (6-14), class sizes (4-11), as well as many different skill levels and interests. Communicated with parents and supervisors to meet goals while providing an engaging class environment. Responsible for multiple classroom locations with different stakeholders, resources, and lesson formats.",
+          "technologies": [
+            "Python",
+            "Scratch",
+            "Blender"
+          ],
+          "createdAt": "2025-03-06T08:17:28.679Z",
+          "updatedAt": "2025-03-06T08:17:28.679Z",
+          "id": "67c95a180e528355496e4451"
+        },
+        {
+          "company": "Scale API",
+          "titles": [
+            {
+              "title": "QA Ops Associate",
+              "id": "67c95a2ef3c2fbef53f39d4b"
+            },
+            {
+              "title": "QA Contractor",
+              "id": "67c95a39f3c2fbef53f39d4d"
+            }
+          ],
+          "dates": {
+            "startDate": "2017",
+            "endDate": "2018",
+            "shortLabel": "2018"
+          },
+          "description": "Ensured superior quality for AI training data annotations through detailed review and feedback to annotators. Promoted to Associate for dedicated effort to the team and projects. Addressed/aided in resolving bugs in annotation software for better annotation quality and annotation ease. Developed training material to improve annotation accuracy and train new annotation reviewers at scale.",
+          "technologies": [
+            "Scale 3D Lidar Annotations",
+            "Google Sheets"
+          ],
+          "createdAt": "2025-03-06T08:19:28.751Z",
+          "updatedAt": "2025-03-06T08:19:28.751Z",
+          "id": "67c95a900e528355496e447f"
+        }
+      ]
     },
     "id": "67c94ade0e528355496e41af"
   }
