--- frontend/src/data/caseStudies/platinumAdvisors.md
@@ -1,29 +1,0 @@
-Platinum Advisors is a national premier advocacy firm. They were looking to redesign their website which highlighted a single state focus to show their expanded national reach while updating the design to appear more modern and premier.
-
-### The Project
-
-I worked together with two main points of contact who were preparing for a release to key stakeholders including the owner of the company. They provided text content as well as a low-fidelity wireframe (in powerpoint) as well as branding and some imagery.
-
-### My Role
-
-I was the sole member from our design studio working on this project. I acted as project manager, designer, and custom developer.
-
-### Challenges:
-
-#### Project Management:
-
-This project was unique in that key stakeholders were not present for any of the meetings and instead and two different representatives that I was in contact with. This was uniquely challenging when paired with an hourly pricing structure that was better suited to the original scope of the project which included on small edits to the project. With an ever expanding scope came a regular need for approval from both stakeholders for budget and revisions. . In order to to meet changes on a fixed deadline I needed to work together with my two points of contact to get unified feedback and approval in a timely fashion. The final project resulted in positive feedback from both points of contact and approval from the key stakeholders at the end with minimal changes despite trade offs made during the process.
-
-#### Design:
-
-The core pieces of the design were selection of stock imagery/video and organizing large amounts of information in an easy to digest way.
-
-_Visual storytelling piece_
-
-_A couple views of the stock imagery on the website, talking about the process of demoing and selection_
-
-_A couple images of the site talking about layout_
-
-#### Custom Development:
-
-I suggested different options for custom styling and plugins to achieve the presentation goals of Platinum Advisors extending Squarespace’s capabilities. We eventually opted for using the Will Meyers Slideshow section plugin, which I custom configured and custom styled with css. To achieve a more premium style on the about page we also opted for come custom coding for portrait frames.
