+++ frontend/src/main.jsx
@@ -1,14 +1,18 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
+// External Imports
+import React from 'react';
+import ReactDOM from 'react-dom/client';
 import { ThemeProvider } from 'styled-components';
-import Theme from './Theme.js'
 
+// Internal Imports
+// App/Router
+import App from './App.jsx';
+// Styles
+import Theme from '@/Theme.js';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <ThemeProvider theme={Theme}>
       <App />
     </ThemeProvider>
-  </React.StrictMode>,
-)
+  </React.StrictMode>
+);
