--- frontend/src/pages/Portfolio.jsx
@@ -1,11 +1,0 @@
-
-
-function Portfolio() {
-    return (
-        <div>
-            <h1>Portfolio</h1>
-        </div>
-    )
-}
-
-export default Portfolio;
